WIP: Deserialize to ADTs
It would be nice to deserialize directly to classes like InputType
rather than having the configuration accept strings and having to
parse them explicitly in the code when they're used.

This is a quasi-working implementation of deserializing directly to
types. The problem is that with this implementation, we have to
pass values as quoted strings like `--input="file"`.

This looks like a detail of Beam's configuration logic, in particular
the following line:

https://github.com/apache/beam/blob/v2.6.0/sdks/java/core/src/main/java/org/apache/beam/sdk/options/PipelineOptionsFactory.java#L1583

That line calls `ObjectMapper::readValue` which expects a JSON token
and throws an IllegalArgumentException if it sees a non-quoted string.

What we need is for the machinery to call `ObjectMapper::convertValue`
which is what it does for simple types like String.

I can't think of an immediate workaround, but welcome suggestions.
I may see if I can put together a PR for beam to allow this in the future.