Use TestPipeline.newProvider() in tests
We've seen some cases in the past where a change deployed to Dataflow
leads to "Value only available at runtime" errors because tests are using
a StaticValueProvider that's available at compile time whereas real
execution doesn't have the value available.

TestPipeline.newProvider() exists for this purpose. You pass in a literal
value, but it won't be available for access until runtime, better simulating
the real execution context.

This change removes usage of StaticValueProvider for all tests where a
TestPipeline is available.