Add PubsubConstraints for attribute truncation (#316)
Mostly addresses #313.

We were seeing the Decoder job fail to progress due to exception being thrown
when attempting to publish error messages to Pubsub. The stack_trace attribute
values were too large.

This change proactively truncates the stack trace values, but then also applies
a final TruncateAttributes transform before attempting to publish. That
introduces the possibility of some data loss in the published attributes, but
seems better than unexpectedly having the pipeline fail.

We are still vulnerable to the possibility of putting too many attributes on
messages (over 100) or having too large of a payload, but we have not yet
observed those cases in practice.