Don't catch Throwable in MapElementsWithErrors
In testing the Decoder, I've seen some records in the error collection due to
OutOfMemoryError. We likely want the worker to fail in this case rather than
send the message to error output.

In general, it's a bit dangerous to catch Throwable since system Errors
(which don't inherit from exception) are generally due to some unrecoverable
state that means the process needs to go down quickly.