Bug: ensure transience and proper projectId for table listing
When this sink was deployed, the default BQ client wasn't finding
any of the tables because they're not in the same project as where
the Dataflow job is running.

We now create BQ clients as needed, and specifically set the project.
We also list all tables for a dataset at once for fewer calls out to BigQuery.

I've deployed this change and it's now working as intended.