Improve the error collection interface
See the new WithErrors class for the main new concept here.
The transforms that emit errors now return a result that has an errorsTo()
method that makes their usage in the top-level Sink and Decoder pipelines
simpler.

This also closes #386 as it refactors the Deduplicate.java code to use the
new error interface.

All the logic for our inputs and outputs have been factored into an io module
and also now use the new error interface consistently.

These refactorings also aid in how the pipeline is displayed in the dataflow UI
so I've removed most of the explicit string names as we apply transforms
since the class names are now instructive. Thus also closes #241.