ingestion-beam style cleanup
A variety of changes to be more consistent in our style internally and more
consistent with some conventions of Beam.

We give all "user-facing" transforms private constructors and public static
factory methods, returning a Singleton instance in case of transforms that
take no parameters.

Also more consistent in skipping explicit names for transforms when they
simply repeat the class name.