Pull schemas from GCS rather than building into jar
Closes #398

This new approach is consistent with how we pull in the geo database and how
we run tests for it. Instead of pulling from GitHub at maven build time,
developers are instructed to run a script to pull down an archive locally.
Then, we rely on Beam's Filesystems library to transparently handle a local
or GCS path.