Metrics per docType
Closes #379

This PR defines some machinery for tracking counters per docType and schema
version using Beam's built-in Metrics interface, which Dataflow automatically
visualizes in its UI and sends to Stackdriver.

This explicitly does not match the segmentation of the AWS-based pipeline.
If we segment on appname, channel, etc. then we quickly end up with counters
numbers in tens of thousands, which Dataflow starts to drop. That number of
user-defined counters also clutters the Dataflow UI and causes it to slow down
significantly.

This also doesn't cover all of the metric types of the old pipeline.
In particular, we aren't tracking compressed bytes and invalid gzip errors.
We could do so in a future PR if these Stackdriver-based metrics prove useful
and worth further investment.