Support outputNumShards and errorNumShards as runtime params
Closes #435

BQ doesn't accept a ValueProvider for its sharding, so we break that out to
a separate compile-time param that we don't expect we'll need to mess with.

We lose some validation capability with this route, but we are now able to
provide a non-zero default for sharding, since we check whether we're streaming
and ignore sharding params for batch mode.