Initialize a new Redis client on connection errors
Closes #451.

When our Redis cluster experienced a failover, we saw that our client got into
a bad state and continually returned connection exceptions. This PR attempts
to handle that case by letting the current message through and initializing a
new client.

This is somewhat of a speculative change, but we should be able to deploy and
verify that it does no harm. We'll see how it performs when Redis next fails
over.