Documentation for refactor with new Republisher job
The Republisher job proposed in this PR would factor out MarkAsSeen
from Decoder which would lead to a more logical flow of data.
It also allows us to share the expense of the MarkAsSeen read with
the read needed to inspect message contents and republish to
smaller topics.

Specifically, the Republisher for structured ingestion would check for
the new debug header and publish message containing that header to
a debug topic per the Glean request in #458.

The Republisher for telemetry data would randomly sample messages to
produce the monitoring topics discussed in #396.