Apply payload transformations to match BQ schema
We support a variety of transformations to account for limitations of BigQuery
schemas. In particular, we support transforming maps into arrays of key/value
pairs, serializing structures to json strings where a string is expected,
and sanitizing field names that BQ cannot support.

We additionally build up a JSON string of items for which we have no
JSON fields in the destination BigQuery table.