Save the uri attribute in nested metadata
After deploying #542, the BigQuery sink jobs were failing. The cause
was that we were writing a "uri" string to the metadata object as extra
identifying information, but that clobbered the "uri" struct of the same name,
leading to a string where BQ expected a record.

We now save the uri string inside the uri struct.