Patch Beam to allow writing into clustered tables (#672)
This is a brittle change that could break if we don't correctly update our
patch when integrating new versions of Beam or if we end up needing to support
writing into both clustered and unclustered tables, so care is needed.

The purpose of this change is to allow us to create our production BigQuery
tables with clustering, which is currently blocked on Beam's lack of support
for configuring clustering. It will be far easier to make that change as
part of setting up production stacks rather than recreating the tables with
clustering after the fact, which will require careful handling to avoid data
loss.

If we integrate this change, we should commit to taking action on helping to
get clustering merged into upstream Beam in its next release.