Add patch for clustering with many writers
BigQueryIO does a kind of weird thing where it spills over writers and
reshuffles data if there are lots of destination tables or high volume of data.
Turns out I missed a case of passing the clustering configuration on that only
crops up in this case where data is reshuffled.

We replace the underlying file from Beam to get the change immediately,
but I will PR the same to upstream Beam so we can remove this after the next
beam release.