Support sample_id for activity-stream pings
Currently, all activity-stream pings are being assigned `sample_id=51`
because they all contain `"client_id":"n/a"`. The pipeline currently discovers
that value, loads it as an attributes, and happily calculates a sample_id from
it.

The change here involves requiring that we have a valid UUID for clientIds
and then also supporting the `impression_id` field in activity-stream pings
that actually does contain a UUID representing the client.