Upgrade to beam 2.17.0
This version includes a fix that we were previously including a patched file
for, so we remove DynamicDestinationsHelpers.java from this repo now.

We still need the patch for WritePartition.java, but the content of that file
did not change in 2.17, so no updates are needed there.

There are also no changes needed for jackson, google cloud SDKs, etc. We are
already pulling in the same versions as used in 2.17.