Draft docs v2 for Account Ecosystem Telemetry
This is an updated design compared to #1109

We now expect ecosystem_anon_id values in the payload rather than as headers.
We have also determined that it's feasible to route a variety of pings to
the AET-specific topics at the edge, so we should be able to use the desktop
telemetry and Glean libraries without modification. More to come on specific
schemas.