[CONSVC-1636] Support glean based telemetry in ContextualServicesReporter
In order to support Glean Telemetry from mobile tiles I created a new
AutoValue class called SponsoredInteraction to contain all the values
necessary for reporting the URLS to our partners.

This required a refactoring of all downstream PTransforms to use this
new class in place of PubSubMessages. I believe that this will
ultimately make it easier to understand the state of the data through
the pipeline.

Test updates are coming in the next commit.