FXA-7921 - Add support for ingesting Cloud Logging messages
This will allow to ingest telemetry from server-side environments shipped to Pub/Sub via Cloud Logging.

ParseLogEntry transform is used to convert LogEntries to format compatible with structured ingestion. This transform was added back in the day for AET and seems like it was intended to be removed in https://github.com/mozilla/gcp-ingestion/pull/1599. I'm repurposing it since it is no longer used.