Relax URL validation rules for suggest clicks (#2614)
We're still getting suggest clicks with the old URL format. These are
likely coming from clients who haven't picked up Remote Settings
changes.

While we investigate what is going on with Remote Settings, we can relax
the validation for non-Amazon advertisers to be a subset of the query
parameters for the old URL format, so that we're not dropping clicks due
to missing parameters.